feat: apply new animation to main page
