+++ src/components/Model/index.js
@@ -5,7 +5,7 @@ import { useFrame } from "@react-three/fiber";
 
 const color = new THREE.Color();
 
-export default function Model({ scroll, ...props }) {
+export const Model = ({ scroll, ...props }) => {
   const group = useRef();
   const { nodes, materials, animations } = useGLTF("/model.glb");
   const { actions } = useAnimations(animations, group);
@@ -69,6 +69,6 @@ export default function Model({ scroll, ...props }) {
       </group>
     </group>
   );
-}
+};
 
 useGLTF.preload("/model.glb");
