+++ src/components/index.js
@@ -2,3 +2,5 @@ export * from "./Header";
 export * from "./InputApply";
 export * from "./CountdownTimer";
 export * from "./SelectTrack";
+export * from "./Overlay";
+export * from "./Model";
